# backend/docs/openapi-0.4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThermoTrace API",
    "description": "API do ThermoTrace — monitoramento térmico de remessas por etiqueta NFC.\n\nVersão 0.4: acesso, perfis, cargas e recepção idempotente de evidências. Declarações de início/checkpoint/final preservadas com cadeia de integridade. Receber uma evidência não confirma o hardware nem libera uma carga. Fila Android e decodificação no servidor estão em implementação.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/remessas": {
      "post": {
        "tags": [
          "Cargas e documentos"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_remessas_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 160,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovaRemessa"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalheRemessa"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Cargas e documentos"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_remessas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatusRemessa"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemessaSaida"
                  },
                  "title": "Response Listar Api V1 Remessas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/remessas/localizar": {
      "get": {
        "tags": [
          "Cargas e documentos"
        ],
        "summary": "Localizar",
        "operationId": "localizar_api_v1_remessas_localizar_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "valor",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 96,
              "title": "Valor"
            }
          },
          {
            "name": "incluir_canceladas",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Incluir Canceladas"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limite"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalizacaoRemessa"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/remessas/{remessa_id}": {
      "get": {
        "tags": [
          "Cargas e documentos"
        ],
        "summary": "Obter",
        "operationId": "obter_api_v1_remessas__remessa_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "remessa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Remessa Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalheRemessa"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/remessas/{remessa_id}/documentos": {
      "post": {
        "tags": [
          "Cargas e documentos"
        ],
        "summary": "Documento",
        "operationId": "documento_api_v1_remessas__remessa_id__documentos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "remessa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Remessa Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 160,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovoDocumento"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentoSaida"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/remessas/{remessa_id}/cancelar": {
      "post": {
        "tags": [
          "Cargas e documentos"
        ],
        "summary": "Cancelar",
        "operationId": "cancelar_api_v1_remessas__remessa_id__cancelar_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "remessa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Remessa Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 160,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MotivoAcao"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemessaSaida"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DetalheRemessa": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "status": {
            "$ref": "#/components/schemas/StatusRemessa"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "empresa_id": {
            "type": "string",
            "format": "uuid",
            "title": "Empresa Id"
          },
          "destinatario_nome": {
            "type": "string",
            "title": "Destinatario Nome"
          },
          "produto_configuracao_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Produto Configuracao Id"
          },
          "perfil_termico_id": {
            "type": "string",
            "format": "uuid",
            "title": "Perfil Termico Id"
          },
          "intervalo_segundos": {
            "type": "integer",
            "title": "Intervalo Segundos"
          },
          "criterio": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criterio"
          },
          "descricao_carga": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao Carga"
          },
          "documentos": {
            "items": {
              "$ref": "#/components/schemas/DocumentoSaida"
            },
            "type": "array",
            "title": "Documentos"
          },
          "volumes": {
            "items": {
              "$ref": "#/components/schemas/VolumeSaida"
            },
            "type": "array",
            "title": "Volumes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "status",
          "criado_em",
          "empresa_id",
          "destinatario_nome",
          "produto_configuracao_id",
          "perfil_termico_id",
          "intervalo_segundos",
          "criterio",
          "descricao_carga",
          "documentos",
          "volumes"
        ],
        "title": "DetalheRemessa"
      },
      "DocumentoSaida": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoDocumento"
          },
          "numero": {
            "type": "string",
            "title": "Numero"
          },
          "serie": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serie"
          },
          "chave_acesso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chave Acesso"
          },
          "cnpj_emitente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj Emitente"
          },
          "conteudo_bruto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conteudo Bruto"
          },
          "digitado_manualmente": {
            "type": "boolean",
            "title": "Digitado Manualmente"
          },
          "validado": {
            "type": "boolean",
            "title": "Validado"
          },
          "observacao_validacao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacao Validacao"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tipo",
          "numero",
          "serie",
          "chave_acesso",
          "cnpj_emitente",
          "conteudo_bruto",
          "digitado_manualmente",
          "validado",
          "observacao_validacao"
        ],
        "title": "DocumentoSaida"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocalizacaoRemessa": {
        "properties": {
          "candidatas": {
            "items": {
              "$ref": "#/components/schemas/RemessaSaida"
            },
            "type": "array",
            "title": "Candidatas"
          },
          "ha_mais": {
            "type": "boolean",
            "title": "Ha Mais"
          }
        },
        "type": "object",
        "required": [
          "candidatas",
          "ha_mais"
        ],
        "title": "LocalizacaoRemessa"
      },
...
      "MotivoAcao": {
        "properties": {
          "motivo": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Motivo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "motivo"
        ],
        "title": "MotivoAcao"
      },
      "NovaRemessa": {
        "properties": {
          "codigo": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._/-]{0,63}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo"
          },
          "produto_configuracao_id": {
            "type": "string",
            "format": "uuid",
            "title": "Produto Configuracao Id"
          },
          "destinatario_nome": {
            "type": "string",
            "maxLength": 240,
            "minLength": 2,
            "title": "Destinatario Nome"
          },
          "descricao_carga": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao Carga"
          },
          "quantidade_volumes": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Quantidade Volumes",
            "default": 1
          },
          "documentos": {
            "items": {
              "$ref": "#/components/schemas/NovoDocumento"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Documentos"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "produto_configuracao_id",
          "destinatario_nome"
        ],
        "title": "NovaRemessa"
      },
...
      "NovoDocumento": {
        "properties": {
          "tipo": {
            "$ref": "#/components/schemas/TipoDocumento"
          },
          "numero": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Numero"
          },
          "chave_acesso": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z0-9]{44}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chave Acesso"
          },
          "serie": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Serie"
          },
          "cnpj_emitente": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z0-9]{12}[0-9]{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj Emitente"
          },
          "conteudo_bruto": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8192
              },
              {
                "type": "null"
              }
            ],
            "title": "Conteudo Bruto"
          },
          "digitado_manualmente": {
            "type": "boolean",
            "title": "Digitado Manualmente",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tipo",
          "numero"
        ],
        "title": "NovoDocumento"
      },
...
      "RemessaSaida": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "status": {
            "$ref": "#/components/schemas/StatusRemessa"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "empresa_id": {
            "type": "string",
            "format": "uuid",
            "title": "Empresa Id"
          },
          "destinatario_nome": {
            "type": "string",
            "title": "Destinatario Nome"
          },
          "produto_configuracao_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Produto Configuracao Id"
          },
          "perfil_termico_id": {
            "type": "string",
            "format": "uuid",
            "title": "Perfil Termico Id"
          },
          "intervalo_segundos": {
            "type": "integer",
            "title": "Intervalo Segundos"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "status",
          "criado_em",
          "empresa_id",
          "destinatario_nome",
          "produto_configuracao_id",
          "perfil_termico_id",
          "intervalo_segundos"
        ],
        "title": "RemessaSaida"
      },
...
      "StatusRemessa": {
        "type": "string",
        "enum": [
          "preparacao",
          "aguardando_aceite",
          "aguardando_coleta",
          "em_transporte",
          "entregue_aguardando_leitura",
          "concluida",
          "cancelada"
        ],
        "title": "StatusRemessa"
      },
      "StatusVolume": {
        "type": "string",
        "enum": [
          "sem_etiqueta",
          "vinculado",
          "monitorando",
          "encerrado"
        ],
        "title": "StatusVolume"
      },
      "TipoDocumento": {
        "type": "string",
        "enum": [
          "nfe",
          "nfce",
          "cte",
          "cte_os",
          "mdfe",
          "awb",
          "pedido",
          "outro",
          "sem_documento"
        ],
        "title": "TipoDocumento"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VolumeSaida": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "sequencia": {
            "type": "integer",
            "title": "Sequencia"
          },
          "identidade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identidade"
          },
          "status": {
            "$ref": "#/components/schemas/StatusVolume"
          },
          "monitorado": {
            "type": "boolean",
            "title": "Monitorado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sequencia",
          "identidade",
          "status",
          "monitorado"
        ],
        "title": "VolumeSaida"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}